refactor(indexer): clean module and test layout
