+++ apps/indexer/src/lib.rs
@@ -14,21 +14,20 @@ pub mod runtime_config;
 pub mod store;
 
 pub use crate::chain::tool::{
-    BatchReadPlanConfig, BlockReadMode, ChainContracts, ChainReadCapability,
-    ChainReadExecutionPlan, ChainReadExecutionReport, ChainReadFailure, ChainReadFailureKind,
-    ChainReadKey, ChainReadMetadata, ChainReadMethod, ChainReadMetrics, ChainReadPlan,
-    ChainReadPlanBuilder, ChainReadReason, ChainReadRequest, ChainReadResult, ChainReadRetryPolicy,
-    ChainReadValue, ChainTool, MulticallReadGroup, PartialChainReadFailureReport, ReadRequirement,
+    BatchReadPlanConfig, BlockReadMode, ChainContracts, ChainReadExecutionReport, ChainReadFailure,
+    ChainReadFailureKind, ChainReadKey, ChainReadMethod, ChainReadMetrics, ChainReadPlan,
+    ChainReadPlanBuilder, ChainReadReason, ChainReadRequest, ChainReadResult, ChainReadValue,
+    ChainTool, MulticallReadGroup, PartialChainReadFailureReport, ReadRequirement,
 };
 pub use crate::datalens::planner::{
     DaoContractAddresses, DaoLogAddressSource, DaoLogQueryPlan, DaoLogSource, DatalensLogPage,
-    DatalensLogQueryReader, DatalensProvisionalCacheSegment, DatalensProvisionalLogPage,
-    DatalensProvisionalLogQueryReader, DatalensProvisionalLogQueryResult, fetch_dao_log_pages,
-    fetch_provisional_dao_log_pages, plan_dao_log_queries,
+    DatalensLogQueryReader, DatalensProvisionalCacheSegment, DatalensProvisionalLogQueryReader,
+    DatalensProvisionalLogQueryResult, fetch_dao_log_pages, fetch_provisional_dao_log_pages,
+    plan_dao_log_queries,
 };
 pub use crate::datalens::warmup::{
-    DatalensWarmupConfig, DatalensWarmupEnsureOutcome, DatalensWarmupEnsurer, DatalensWarmupKind,
-    DatalensWarmupSubmitRequest, ensure_datalens_warmup_task, follow_query_request,
+    DatalensWarmupConfig, DatalensWarmupEnsureOutcome, DatalensWarmupEnsurer,
+    DatalensWarmupSubmitRequest, ensure_datalens_warmup_task,
 };
 pub use crate::datalens::{
     DatalensLogQueryCacheOutcome, DatalensLogQueryCacheSummary, DatalensLogQueryResult,
@@ -42,21 +41,20 @@ pub use crate::decode::dao_event::{
     DecodedGovernorEvent, DecodedTimelockEvent, DecodedTokenEvent, DelegateChangedEvent,
     DelegateVotesChangedEvent, GovernanceTokenStandard, ParameterChangeEvent, ProposalCreatedEvent,
     ProposalExtendedEvent, ProposalIdEvent, ProposalQueuedEvent, RoleAccountEvent,
-    RoleAdminChangedEvent, TimelockChangeEvent, TimelockOperationIdEvent, TokenTransferEvent,
-    UnsupportedTopicEvent, VoteCastEvent, VoteCastWithParamsEvent, decode_dao_log,
+    RoleAdminChangedEvent, TimelockOperationIdEvent, TokenTransferEvent, UnsupportedTopicEvent,
+    VoteCastEvent, VoteCastWithParamsEvent, decode_dao_log,
 };
 pub use crate::decode::evm_log::{
     EvmLogNormalizationError, NormalizedEvmLog, normalize_evm_log_rows,
 };
 pub use crate::onchain::refresh::{
     ChainToolOnchainRefreshReader, DEFAULT_ONCHAIN_REFRESH_APPLY_BATCH_SIZE, EvmRpcChainTool,
-    LivePowerOverlayReader, LivePowerOverlayRefreshError, MultiChainToolOnchainRefreshReader,
-    OnchainRefreshReadReport, OnchainRefreshReadValue, OnchainRefreshReader,
-    OnchainRefreshReaderError, OnchainRefreshRunReport, OnchainRefreshTask,
+    LivePowerOverlayReader, MultiChainToolOnchainRefreshReader, OnchainRefreshReadValue,
+    OnchainRefreshReader, OnchainRefreshReaderError, OnchainRefreshRunReport, OnchainRefreshTask,
     OnchainRefreshTaskScope, OnchainRefreshTickClock, OnchainRefreshTickConfig,
     OnchainRefreshTickReport, OnchainRefreshTickRunner, OnchainRefreshTickScheduler,
     OnchainRefreshTickSkipReason, OnchainRefreshWorker, OnchainRefreshWorkerConfig,
-    OnchainRefreshWorkerError, SystemOnchainRefreshTickClock, refresh_live_power_overlays,
+    OnchainRefreshWorkerError, refresh_live_power_overlays,
 };
 pub use crate::projection::data_metric::DataMetricWrite;
 pub use crate::projection::power_reconcile::{
@@ -66,47 +64,44 @@ pub use crate::projection::power_reconcile::{
 };
 pub use crate::projection::proposal::{
     InMemoryProposalProjectionRepository, ProposalActionWrite, ProposalCreatedWrite,
-    ProposalDeadlineExtensionWrite, ProposalEventCommon, ProposalExtendedWrite, ProposalIdWrite,
+    ProposalDeadlineExtensionWrite, ProposalExtendedWrite, ProposalIdWrite,
     ProposalProjectionBatch, ProposalProjectionContext, ProposalProjectionError,
     ProposalProjectionEvent, ProposalProjectionRepository, ProposalQueuedWrite,
-    ProposalRepositoryWriteError, ProposalStateEpochWrite, ProposalStateWriteKind,
-    ProposalTimestampBackfillCandidate, ProposalTimestampBackfillUpdate, ProposalWrite,
-    plan_proposal_timestamp_backfill_updates, project_proposal_events,
+    ProposalStateEpochWrite, ProposalStateWriteKind, ProposalTimestampBackfillCandidate,
+    ProposalTimestampBackfillUpdate, ProposalWrite, plan_proposal_timestamp_backfill_updates,
+    project_proposal_events,
 };
 pub use crate::projection::proposal_metadata::{
     ProposalTextMetadata, ProposalTitleExtractionError, ProposalTitleExtractor,
     derive_proposal_metadata, derive_proposal_metadata_with_title_extractor,
 };
 pub use crate::projection::timelock::{
-    InMemoryTimelockProjectionRepository, TIMELOCK_POSTGRES_ADAPTER_GAP, TimelockCallWrite,
-    TimelockEventCommon, TimelockMinDelayChangeWrite, TimelockOperationHintWrite,
-    TimelockOperationWrite, TimelockProjectionBatch, TimelockProjectionContext,
-    TimelockProjectionError, TimelockProjectionEvent, TimelockProjectionRepository,
-    TimelockProposalActionLink, TimelockProposalLinkContext, TimelockRepositoryWriteError,
+    InMemoryTimelockProjectionRepository, TimelockCallWrite, TimelockMinDelayChangeWrite,
+    TimelockOperationHintWrite, TimelockOperationWrite, TimelockProjectionBatch,
+    TimelockProjectionContext, TimelockProjectionError, TimelockProjectionEvent,
+    TimelockProjectionRepository, TimelockProposalActionLink, TimelockProposalLinkContext,
     TimelockRoleEventWrite, project_timelock_events, project_timelock_events_with_proposal_links,
 };
 pub use crate::projection::token::{
-    ContributorWrite, DataMetricTokenDelta, DelegateChangedWrite, DelegateMappingWrite,
-    DelegateRollingWrite, DelegateVotesChangedWrite, DelegateWrite,
+    DelegateChangedWrite, DelegateRollingWrite, DelegateVotesChangedWrite,
     InMemoryTokenProjectionRepository, TokenEventCommon, TokenProjectionBatch,
     TokenProjectionContext, TokenProjectionError, TokenProjectionEvent, TokenProjectionOperation,
-    TokenProjectionRepository, TokenRepositoryWriteError, TokenTransferWrite, project_token_events,
+    TokenProjectionRepository, TokenTransferWrite, project_token_events,
 };
 pub use crate::projection::vote::{
-    ContributorVoteSignalWrite, DataMetricVoteDelta, InMemoryVoteProjectionRepository,
-    ProposalVoteTotalWrite, VoteCastGroupWrite, VoteCastWithParamsWrite, VoteCastWrite,
-    VoteEventCommon, VoteProjectionBatch, VoteProjectionContext, VoteProjectionError,
-    VoteProjectionEvent, VoteProjectionRepository, VoteRepositoryWriteError, project_vote_events,
+    ContributorVoteSignalWrite, InMemoryVoteProjectionRepository, ProposalVoteTotalWrite,
+    VoteCastGroupWrite, VoteCastWithParamsWrite, VoteCastWrite, VoteProjectionBatch,
+    VoteProjectionContext, VoteProjectionError, VoteProjectionEvent, VoteProjectionRepository,
+    project_vote_events,
 };
 pub use crate::store::postgres::{
     DEFAULT_ONCHAIN_REFRESH_DEFERRED_DRAIN_ROWS, PostgresIndexerRunnerStore,
-    PostgresIndexerRunnerStoreError, PostgresIndexerRunnerTransaction,
     PostgresProvisionalCleanupStore, PostgresProvisionalPowerOverlayStore,
     PostgresProvisionalProposalOverlayStore, PostgresProvisionalSegmentStore,
     ProposalReferenceFieldCandidate, ProposalReferenceFieldUpdate, ProposalTitleRefreshCandidate,
     ProposalTitleRefreshUpdate, read_proposal_reference_field_candidates,
-    read_proposal_timestamp_backfill_candidates, read_proposal_title_refresh_candidates,
-    update_proposal_reference_fields, update_proposal_timestamp_backfill, update_proposal_titles,
+    read_proposal_title_refresh_candidates, update_proposal_reference_fields,
+    update_proposal_titles,
 };
 pub use checkpoint::{
     CheckpointBlockRange, CheckpointRepository, IndexerCheckpoint, IndexerCheckpointIdentity,
@@ -121,8 +116,7 @@ pub use datalens::{
     DatalensDurableHeadReader, DatalensNativeClient, DatalensNativeReader, ServiceReadiness,
     verify_datalens_service,
 };
-pub use error::{CheckpointError, ConfigError, DatalensError, IndexerError};
-pub use graphql::IndexerGraphqlSchema;
+pub use error::{CheckpointError, ConfigError, DatalensError};
 pub use provisional::{
     DatalensProvisionalSegmentStore, DatalensProvisionalSegmentWrite, ProvisionalCleanupReport,
     ProvisionalCleanupStore, ProvisionalContributorPowerOverlayWrite,
@@ -130,22 +124,20 @@ pub use provisional::{
     ProvisionalPowerOverlayScope, ProvisionalPowerOverlayStore, ProvisionalProposalOverlayStore,
     ProvisionalProposalOverlayWrite, ProvisionalRollbackReport, ProvisionalRollbackScope,
     ProvisionalSegmentCleanupCandidate, ProvisionalSegmentCleanupDecision,
-    ProvisionalTimelockOperationOverlayWrite, ProvisionalWorker, ProvisionalWorkerError,
-    ProvisionalWorkerOptions, ProvisionalWorkerReport, plan_provisional_segment_cleanup,
+    ProvisionalTimelockOperationOverlayWrite, ProvisionalWorker, ProvisionalWorkerOptions,
+    plan_provisional_segment_cleanup,
 };
 pub use runner::{
-    AdaptiveChunkFeedback, AdaptiveChunkSizer, AdaptiveChunkSizerConfig,
-    AdaptiveChunkSizingDecision, AdaptiveChunkSizingReason, DaoEventDecoder,
-    InMemoryIndexerRunnerStore, InMemoryIndexerRunnerStoreError, IndexerEventDecoder,
-    IndexerOnchainRefreshTick, IndexerProjectionBatch, IndexerRunner, IndexerRunnerContexts,
-    IndexerRunnerError, IndexerRunnerOptions, IndexerRunnerProgress, IndexerRunnerReport,
-    IndexerRunnerStore, IndexerRunnerTransaction, ProposalTimestampBackfillConfig, page_rows,
+    AdaptiveChunkFeedback, AdaptiveChunkSizer, AdaptiveChunkSizerConfig, AdaptiveChunkSizingReason,
+    DaoEventDecoder, InMemoryIndexerRunnerStore, IndexerEventDecoder, IndexerOnchainRefreshTick,
+    IndexerProjectionBatch, IndexerRunner, IndexerRunnerContexts, IndexerRunnerOptions,
+    IndexerRunnerReport, IndexerRunnerStore, IndexerRunnerTransaction,
+    ProposalTimestampBackfillConfig, page_rows,
 };
 pub use runtime_config::{
-    AdaptiveChunkSizerRuntimeConfig, ContractSetConcurrencyLimit, GraphqlRuntimeConfig,
-    IndexerContractSetMode, IndexerContractSetRuntimeConfig, IndexerRuntimeConfig,
-    IndexerTargetHeight, OnchainRefreshRpcChainConfig, OnchainRefreshRuntimeConfig,
-    ProvisionalRuntimeConfig, datalens_retry_config, onchain_refresh_apply_batch_size_from_env,
-    onchain_refresh_debounce_from_env, onchain_refresh_deferred_drain_batch_size_from_env,
-    onchain_refresh_worker_enabled, parse_bool_env_value, parse_i64_env_value, required_env,
+    ContractSetConcurrencyLimit, GraphqlRuntimeConfig, IndexerContractSetMode,
+    IndexerContractSetRuntimeConfig, IndexerRuntimeConfig, IndexerTargetHeight,
+    OnchainRefreshRuntimeConfig, ProvisionalRuntimeConfig, datalens_retry_config,
+    onchain_refresh_debounce_from_env, onchain_refresh_worker_enabled, parse_bool_env_value,
+    parse_i64_env_value, required_env,
 };
