+++ apps/indexer/src/projection/token.rs
@@ -807,196 +807,6 @@ fn is_nonzero_decimal(value: &str) -> bool {
     normalize_decimal(value) != "0"
 }
 
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    const CONTRACT_SET_ID: &str = "demo-contract-set";
-    const DAO_CODE: &str = "demo-dao";
-    const GOVERNOR: &str = "0x00000000000000000000000000000000000000a1";
-    const TOKEN: &str = "0x00000000000000000000000000000000000000a2";
-    const DELEGATOR: &str = "0x0000000000000000000000000000000000000c01";
-    const DELEGATE: &str = "0x0000000000000000000000000000000000000c02";
-    const SECOND_DELEGATE: &str = "0x0000000000000000000000000000000000000c03";
-    const RECEIVER: &str = "0x0000000000000000000000000000000000000c04";
-    const ZERO_ADDRESS: &str = "0x0000000000000000000000000000000000000000";
-
-    #[test]
-    fn test_redelegate_marks_previous_relation_inactive_with_zero_power() {
-        let repository = project_events(vec![
-            delegate_changed("delegate", 1, 0, 1, ZERO_ADDRESS, DELEGATE),
-            transfer("mint", 2, 0, 1, ZERO_ADDRESS, DELEGATOR, "100"),
-            delegate_changed("redelegate", 3, 0, 1, DELEGATE, SECOND_DELEGATE),
-        ]);
-
-        let previous_relation = repository
-            .delegates()
-            .get(&delegate_ref(DELEGATOR, DELEGATE))
-            .expect("previous relation should be staged");
-        assert!(!previous_relation.is_current);
-        assert_eq!(previous_relation.power, "0");
-
-        let current_relation = repository
-            .delegates()
-            .get(&delegate_ref(DELEGATOR, SECOND_DELEGATE))
-            .expect("current relation should be staged");
-        assert!(current_relation.is_current);
-
-        let mapping = repository
-            .delegate_mappings()
-            .get(DELEGATOR)
-            .expect("delegate mapping should be staged");
-        assert_eq!(mapping.to, SECOND_DELEGATE);
-    }
-
-    #[test]
-    fn test_power_update_preserves_delegate_mapping_relation_metadata() {
-        let repository = project_events(vec![
-            delegate_changed("delegate", 1, 0, 1, ZERO_ADDRESS, DELEGATE),
-            transfer("mint", 2, 0, 1, ZERO_ADDRESS, DELEGATOR, "40"),
-        ]);
-
-        let mapping = repository
-            .delegate_mappings()
-            .get(DELEGATOR)
-            .expect("delegate mapping should be staged");
-        assert_eq!(mapping.to, DELEGATE);
-        assert_eq!(mapping.power, "40");
-        assert_eq!(mapping.common.block_number, "1");
-        assert_eq!(mapping.common.transaction_hash, "0xtx10");
-    }
-
-    #[test]
-    fn test_redelegated_power_update_preserves_current_relation_metadata() {
-        let repository = project_events(vec![
-            delegate_changed("delegate", 1, 0, 1, ZERO_ADDRESS, DELEGATE),
-            transfer("mint", 1, 0, 2, ZERO_ADDRESS, DELEGATOR, "100"),
-            delegate_changed("redelegate", 2, 0, 1, DELEGATE, SECOND_DELEGATE),
-            delegate_votes_changed("second-votes", 2, 0, 2, SECOND_DELEGATE, "0", "100"),
-            transfer("send", 3, 0, 1, DELEGATOR, RECEIVER, "25"),
-        ]);
-
-        let mapping = repository
-            .delegate_mappings()
-            .get(DELEGATOR)
-            .expect("delegate mapping should be staged");
-        assert_eq!(mapping.to, SECOND_DELEGATE);
-        assert_eq!(mapping.power, "75");
-        assert_eq!(mapping.common.block_number, "2");
-        assert_eq!(mapping.common.transaction_hash, "0xtx20");
-    }
-
-    fn project_events(events: Vec<TokenProjectionEvent>) -> InMemoryTokenProjectionRepository {
-        let batch = project_token_events(&token_projection_context(), events)
-            .expect("token projection should succeed");
-        let mut repository = InMemoryTokenProjectionRepository::default();
-        repository
-            .apply(&batch)
-            .expect("in-memory token writes should succeed");
-        repository
-    }
-
-    fn delegate_changed(
-        id: &str,
-        block_number: u64,
-        transaction_index: u64,
-        log_index: u64,
-        from_delegate: &str,
-        to_delegate: &str,
-    ) -> TokenProjectionEvent {
-        TokenProjectionEvent {
-            log: normalized_log(id, block_number, transaction_index, log_index),
-            event: DecodedTokenEvent::DelegateChanged(DelegateChangedEvent {
-                delegator: DELEGATOR.to_owned(),
-                from_delegate: from_delegate.to_owned(),
-                to_delegate: to_delegate.to_owned(),
-            }),
-        }
-    }
-
-    fn delegate_votes_changed(
-        id: &str,
-        block_number: u64,
-        transaction_index: u64,
-        log_index: u64,
-        delegate: &str,
-        previous_votes: &str,
-        new_votes: &str,
-    ) -> TokenProjectionEvent {
-        TokenProjectionEvent {
-            log: normalized_log(id, block_number, transaction_index, log_index),
-            event: DecodedTokenEvent::DelegateVotesChanged(DelegateVotesChangedEvent {
-                delegate: delegate.to_owned(),
-                previous_votes: previous_votes.to_owned(),
-                new_votes: new_votes.to_owned(),
-            }),
-        }
-    }
-
-    fn transfer(
-        id: &str,
-        block_number: u64,
-        transaction_index: u64,
-        log_index: u64,
-        from: &str,
-        to: &str,
-        value: &str,
-    ) -> TokenProjectionEvent {
-        TokenProjectionEvent {
-            log: normalized_log(id, block_number, transaction_index, log_index),
-            event: DecodedTokenEvent::Transfer(TokenTransferEvent {
-                from: from.to_owned(),
-                to: to.to_owned(),
-                value: value.to_owned(),
-                standard: GovernanceTokenStandard::Erc20,
-            }),
-        }
-    }
-
-    fn token_projection_context() -> TokenProjectionContext {
-        TokenProjectionContext {
-            contract_set_id: CONTRACT_SET_ID.to_owned(),
-            dao_code: DAO_CODE.to_owned(),
-            governor_address: GOVERNOR.to_owned(),
-            token_address: TOKEN.to_owned(),
-            contracts: ChainContracts {
-                governor: GOVERNOR.to_owned(),
-                governor_token: TOKEN.to_owned(),
-                timelock: String::new(),
-            },
-            token_standard: GovernanceTokenStandard::Erc20,
-            from_block: 1,
-            to_block: 100,
-            target_height: None,
-            read_plan_config: BatchReadPlanConfig::default(),
-            current_power_method: ChainReadMethod::GetVotes,
-        }
-    }
-
-    fn normalized_log(
-        id: &str,
-        block_number: u64,
-        transaction_index: u64,
-        log_index: u64,
-    ) -> NormalizedEvmLog {
-        NormalizedEvmLog {
-            id: id.to_owned(),
-            chain_id: 1,
-            block_number,
-            block_hash: format!("0xblock{block_number}"),
-            block_timestamp_ms: Some((1_700_000_000 + block_number) * 1_000),
-            transaction_hash: format!("0xtx{block_number}{transaction_index}"),
-            transaction_index,
-            log_index,
-            address: TOKEN.to_owned(),
-            topics: vec![],
-            data: "0x".to_owned(),
-            removed: false,
-            raw_payload: serde_json::json!({ "block_number": block_number }),
-        }
-    }
-}
-
 fn is_negative_decimal(value: &str) -> bool {
     value.starts_with('-') && is_nonzero_decimal(value.trim_start_matches('-'))
 }
